all admin handlers need to be in routes to work
