+++ aries_cloudagent/protocols/coordinate_mediation/v1_0/routes.py
@@ -11,7 +11,7 @@ from aiohttp_apispec import (
     response_schema,
 )
 
-from marshmallow import fields
+from marshmallow import fields, validate
 
 from .models.mediation_record import (MediationRecord,
                                       MediationRecordSchema,
@@ -27,8 +27,6 @@ from .messages.mediate_request import MediationRequest, MediationRequestSchema
 from .messages.mediate_grant import MediationGrantSchema
 from .messages.mediate_deny import MediationDenySchema
 from .manager import MediationManager as M_Manager
-from .keylist_routes import (send_keylists_request, list_all_records,
-                             send_update_keylists, update_keylists)
 from ....messaging.models.base import BaseModelError
 from ....messaging.models.openapi import OpenAPISchema
 
@@ -39,6 +37,13 @@ from ....utils.tracing import trace_event, get_timer
 
 from .message_types import SPEC_URI
 from operator import itemgetter
+from .keylist_routes import (send_keylists_request, list_all_records,
+                             send_update_keylists, update_keylists)
+from aries_cloudagent.protocols.coordinate_mediation.v1_0.messages.inner.keylist_update_rule import KeylistUpdateRule
+from aries_cloudagent.protocols.coordinate_mediation.v1_0.manager import MediationManager
+from aries_cloudagent.protocols.coordinate_mediation.v1_0.messages.keylist_query import KeylistQuery
+from aries_cloudagent.protocols.routing.v1_0.models.route_record import RouteRecord, RouteRecordSchema
+from .messages.keylist_update_response import KeylistUpdateResponseSchema
 
 
 class MediationListSchema(OpenAPISchema):
@@ -506,6 +511,221 @@ async def mediation_record_deny(request: web.BaseRequest):
                 outcome="mediation_deny.END", perf_counter=r_time,
                 )
     return web.json_response(result, status=201)
+
+
+# class AllKeyListRecordsPagingSchema(OpenAPISchema):
+#     """Parameters and validators for keylist record list query string."""
+
+#     #filter = fields..... TODO: add filtering to handler
+#     limit= fields.Integer(
+#         description="Number of keylists in a single page.",
+#         required=False,
+#         example="5",
+#     )
+#     offset= fields.Integer(
+#         description="Page to receive in pagination.",
+#         required=False,
+#         example="5",
+#     )
+
+
+class KeyListRecordListSchema(OpenAPISchema):
+    """Result schema for mediation list query."""
+
+    results = fields.List(  # TODO: order matters, should match sequence?
+        fields.Nested(RouteRecordSchema),
+        description="List of keylist records",
+    )
+
+
+class KeylistUpdateSchema(OpenAPISchema):
+    """Routing key schema."""
+
+    action = fields.Str(
+        description="update actions",
+        required=True,
+        validate=validate.OneOf(
+            [
+                    getattr(KeylistUpdateRule, m)
+                    for m in vars(KeylistUpdateRule)
+                    if m.startswith("RULE_")
+            ]
+        ),
+        example="'add' or 'remove'",
+    )
+    key = fields.Str(
+        description = "Key to be acted on.",
+        required=True,
+    )
+
+
+class KeylistUpdateRequestSchema(OpenAPISchema):
+    """keylist update request schema"""
+    updates = fields.List(
+        fields.Nested(
+            KeylistUpdateSchema()
+        )
+    )
+
+
+class MediationIdMatchInfoSchema(OpenAPISchema):
+    """Path parameters and validators for request taking mediation request id."""
+
+    mediation_id = MEDIATION_ID_SCHEMA
+
+
+@docs(
+    tags=["keylist"],
+    summary="Query keylists, returns list of all keylist records.",
+)
+# @querystring_schema(AllRecordsQueryStringSchema()) # TODO: add filtering
+@response_schema(KeyListRecordListSchema(), 200)
+async def list_all_records(request: web.BaseRequest):
+    """
+    Request handler for searching keylist records.
+
+    Args:
+        request: aiohttp request object
+
+    Returns:
+        keylists
+
+    """
+    context = request.app["request_context"]
+    try:
+        # TODO: use new keylist models
+        records = await RouteRecord.query(context)
+        results = [record.serialize() for record in records]
+    except (StorageError, BaseModelError) as err:
+        raise web.HTTPBadRequest(reason=err.roll_up) from err
+    return web.json_response({"results": results})
+
+
+@docs(
+    tags=["keylist"],
+    summary="send Query keylists request, returns list of all keylist records.",
+)
+# @querystring_schema(AllRecordsQueryStringSchema()) # TODO: add filtering
+@match_info_schema(MediationIdMatchInfoSchema())
+@response_schema(KeyListRecordListSchema(), 200)
+async def send_keylists_request(request: web.BaseRequest):
+    """
+    Request handler for searching keylist records.
+
+    Args:
+        request: aiohttp request object
+
+    Returns:
+        keylists
+
+    """
+    r_time = get_timer()
+    context = request.app["request_context"]
+    outbound_handler = request.app["outbound_message_router"],
+    mediation_id = request.match_info["mediation_id"]
+    # body = await request.json()
+    record = None
+    try:
+        record = await MediationRecord.retrieve_by_id(
+            context, mediation_id
+        )
+        # TODO: add pagination to request
+        request = KeylistQuery()
+    except (StorageError, BaseModelError) as err:
+        raise web.HTTPBadRequest(reason=err.roll_up) from err
+
+    await outbound_handler(
+        request, connection_id=record.connection_id
+    )
+    trace_event(
+        context.settings,
+        request,
+        outcome="keylist_update_request.END",
+        perf_counter=r_time,
+    )
+    return web.json_response({"results": request})
+
+
+@docs(tags=["keylist"], summary="update keylist.")
+@match_info_schema(MediationIdMatchInfoSchema())
+@request_schema(KeylistUpdateRequestSchema())
+@response_schema(KeylistUpdateResponseSchema(), 201)
+async def update_keylists(request: web.BaseRequest):
+    """
+    Request handler for updating keylist.
+
+    Args:
+        request: aiohttp request object
+    """
+    r_time = get_timer()
+    context = request.app["request_context"]
+    mediation_id = request.match_info["mediation_id"]
+    body = await request.json()
+    updates = body.get("updates")
+    record = None
+    try:
+        trace_event(
+            context.settings, record,
+            outcome="keylist_update.START",
+        )
+        record = await MediationRecord.retrieve_by_id(
+            context, mediation_id
+        )
+        if record.state != MediationRecord.STATE_GRANTED:
+            raise web.HTTPBadRequest(reason=("mediation is not granted."))
+        mgr = MediationManager(context)
+        response = await mgr.update_keylist(
+            record, updates=updates
+        )
+    except (StorageError, BaseModelError) as err:
+        raise web.HTTPBadRequest(reason=err.roll_up) from err
+    trace_event(
+        context.settings,
+        response,
+        outcome="keylist_update.END",
+        perf_counter=r_time,
+    )
+    return web.json_response(response, status=201)
+
+
+@docs(tags=["keylist"], summary="update keylist.")
+@match_info_schema(MediationIdMatchInfoSchema())
+#@querystring_schema(KeylistUpdateRequestSchema())
+@request_schema(KeylistUpdateRequestSchema())
+@response_schema(KeylistUpdateResponseSchema(), 201)
+async def send_update_keylists(request: web.BaseRequest):
+    """
+    Request handler for updating keylist.
+
+    Args:
+        request: aiohttp request object
+    """
+    r_time = get_timer()
+    context = request.app["request_context"]
+    outbound_handler = request.app["outbound_message_router"]
+    mediation_id = request.match_info["mediation_id"]
+    body = await request.json()
+    updates = body.get("updates")
+    updates =[KeylistUpdateRule(
+        recipient_key=update.get("key"),
+        action=update.get("action")) for update in updates]
+    try:
+        record = await MediationRecord.retrieve_by_id(
+            context, mediation_id
+        )
+        if record.state != MediationRecord.STATE_GRANTED:
+            raise web.HTTPBadRequest(reason=("mediation is not granted."))
+        manager = MediationManager(context)
+        request = await manager.update_keylist(
+            record=record, updates=updates
+        )
+        results = request.serialize() 
+    except (StorageError, BaseModelError) as err:
+        raise web.HTTPBadRequest(reason=err.roll_up) from err
+    await outbound_handler(
+        request, connection_id=record.connection_id
+    )
+    return web.json_response(results, status=201)
 
 
 async def register(app: web.Application):
